Update create order tests

+++ Libs/RichillCapital.Contracts/Orders/CreateOrderRequest.cs
@@ -7,5 +7,5 @@ public sealed record CreateOrderRequest
     public required string TradeType { get; init; }
     public required string OrderType { get; init; }
     public required string TimeInForce { get; init; }
-    public required int Quantity { get; init; }
+    public required decimal Quantity { get; init; }
 }
