+++ Libs/RichillCapital.UseCases/Orders/Commands/CreateOrderCommand.cs
@@ -12,5 +12,5 @@ public sealed record CreateOrderCommand :
     public required string TradeType { get; init; }
     public required string OrderType { get; init; }
     public required string TimeInForce { get; init; }
-    public required int Quantity { get; init; }
+    public required decimal Quantity { get; init; }
 }
